thymeleafe

+++ src/main/java/ru/top/java212/dao/AddressRepository.java
@@ -4,5 +4,5 @@ package ru.top.java212.dao;
 import org.springframework.data.repository.CrudRepository;
 import ru.top.java212.model.Address;
 
-public interface AddressRepository extends CrudRepository<Address, Long> {
+public interface AddressRepository extends CrudRepository<Address, Integer> {
 }
