+++ src/main/java/ru/top/java212/dao/MyOrdersRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import ru.top.java212.model.MyOrders;
 
-public interface MyOrdersRepository extends JpaRepository<MyOrders, Long> {
+public interface MyOrdersRepository extends JpaRepository<MyOrders, Integer> {
 }
